[bug] Locs_vox must be integers
The calculation of voxel location (Locs_vox) can lead to values that
are not strictly integers due to calculation inaccuracies. To avoid an
error later in the code, we check that the values stored in Locs_vox
are integers at a given level of absolute tolerance and if so, convert
all values in Locs_vox into integers.