Fixed bug where threshold TH is wrongly scaled by GX
A user specified the (rarely used) ANCOVA normalisation, based on
computed global means.  He *did* *not* specify proportional scaling or
global scaling of any sort.  However, on line 504 of snpm_ui, TH was
then scaled by GX… an innocuous error since TH is -Inf, *except* one of
the globals happened to be *negative*, changing the TH to +Inf!

Hopefully this now fixes it.